Add files via upload
This solution determines the number of possible paths, and represents how each node is turned in binary. It then loops through all possible solutions
to determine the best route. Although it finds the correct route every time, it does not scale as well with large triangles as my previous solution does.
I attempted to mediate this by breaking a large triangle into smaller ones and iterating through the large triangle by saving the best possible paths through
the smaller triangles, but this would often generate incorrect answers and was not as fast as the regular Monte Carlo method.